extract AppSideNavCollapseToggle
