+++ src/app/base/components/AppSideNavigation/AppSideNavCollapseToggle/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./AppSideNavCollapseToggle";
