+++ src/app/base/components/AppSideNavigation/_index.scss
@@ -70,7 +70,7 @@
       background: inherit;
     }
     &.is-collapsed {
-      .p-side-navigation__collapse-toggle,
+      .l-navigation-collapse-toggle,
       .l-navigation__controls {
         i {
           transform: rotate(180deg);
